+++ code/gui/models/qmainnavigationmodel.h
@@ -68,7 +68,7 @@ public:
     Q_INVOKABLE QObject* getSubModel(int index) const;
     Q_INVOKABLE int getSubModelType(int index) const;
     Q_INVOKABLE QString getModelTitle() const;
-    Q_INVOKABLE int getSubModelCount(void) const { return m_submodels.size(); }
+    Q_INVOKABLE unsigned int getSubModelCount(void) const { return static_cast<unsigned int>(m_submodels.size()); }
 
     void addSubModel(std::string displayString, QAbstractItemModel *listModel, ModelTypes::SubModelType modelType);
